+++ sdk/asset/index.js
@@ -1,2 +1,0 @@
-const asset = window.canva.content;
-export const upload = asset.upload.bind(asset);