+++ sdk/preview/asset/index.js
@@ -1,4 +1,5 @@
 const asset = window.canva.content;
 export const upload = asset.upload.bind(asset);
-/** beta */
 export const getTemporaryUrl = asset.getTemporaryUrl.bind(asset);
+export const findFonts = asset.findFonts.bind(asset);
+export const requestFontSelection = asset.requestFontSelection.bind(asset);