+++ sdk/user/index.js
@@ -1,5 +1,0 @@
-const a = window.canva.authentication;
-export const auth = Object.freeze({
-  requestAuthentication: a.requestAuthentication.bind(a),
-  getCanvaUserToken: a.getCanvaUserToken.bind(a),
-});